Split result
